fix issue when use is not logged in

+++ packages/event-overview/src/hooks/use-favorites.tsx
@@ -7,7 +7,7 @@ import { useUser } from '../context/user-context';
  */
 export function useFavorites() {
   const { apiUrl, siteId } = useConfig();
-  const { jwt } = useUser();
+  const { jwt, isLoggedIn } = useUser();
   const { data, error, mutate } = useSWR(`/event/favorites`);
 
   /**
@@ -15,17 +15,19 @@ export function useFavorites() {
    * @param eventId
    */
   async function favorite(event: any) {
-    const res = await fetch(
-      `${apiUrl}/api/site/${siteId}/event/${event.id}/favorite`,
-      {
-        method: 'POST',
-        headers: {
-          'X-Authorization': `Bearer ${jwt}`,
-        },
+    if (isLoggedIn()) {
+      const res = await fetch(
+        `${apiUrl}/api/site/${siteId}/event/${event.id}/favorite`,
+        {
+          method: 'POST',
+          headers: {
+            'X-Authorization': `Bearer ${jwt}`,
+          },
+        }
+      );
+      if (res.ok) {
+        mutate(data.concat([event]));
       }
-    );
-    if (res.ok) {
-      mutate(data.concat([event]));
     }
   }
 
@@ -34,22 +36,26 @@ export function useFavorites() {
    * @param eventId
    */
   async function unfavorite(eventId: number) {
-    const res = await fetch(
-      `${apiUrl}/api/site/${siteId}/event/${eventId}/favorite`,
-      {
-        method: 'DELETE',
-        headers: {
-          'X-Authorization': `Bearer ${jwt}`,
-        },
+    if (isLoggedIn()) {
+      const res = await fetch(
+        `${apiUrl}/api/site/${siteId}/event/${eventId}/favorite`,
+        {
+          method: 'DELETE',
+          headers: {
+            'X-Authorization': `Bearer ${jwt}`,
+          },
+        }
+      );
+      if (res.ok) {
+        // remove event from data array
+        mutate([].concat(data).filter((event: any) => event.id !== eventId));
       }
-    );
-    if (res.ok) {
-      // remove event from data array
-      mutate([].concat(data).filter((event: any) => event.id !== eventId));
     }
   }
 
   function isFavorite(eventId: any) {
+    if (!data || !data.length) return false;
+    if (!isLoggedIn()) return false;
     return data.some((event: any) => event.id === eventId);
   }
